dotTranform теперь возвращает Double в соответствии с параметром y

+++ src/main/java/utils/DotTransformer.java
@@ -2,8 +2,8 @@ package utils;
 
 public class DotTransformer {
 
-    public static Float dotTransform(String num){
+    public static Double dotTransform(String num){
         String strNum = num.replace(",",".");
-        return Float.parseFloat(strNum);
+        return Double.parseDouble(strNum);
     }
 }
